Fanasus

+++ app/src/main/java/gr/gdschua/bloodapp/Utils/CacheClearer.java
@@ -16,9 +16,8 @@ public class CacheClearer {
         if (dir != null && dir.isDirectory()) {
             String[] children = dir.list();
             for (int i = 0; i < children.length; i++) {
-                boolean success = deleteDir(new File(dir, children[i]));
-                if (!success) {
-                    return false;
+                if(children[i].startsWith("picker")) {
+                    deleteDir(new File(dir, children[i]));
                 }
             }
             return dir.delete();
